requirements.txt
blinker
Click
decorator
Flask
Flask-Login
Flask-Mail
Flask-pymysql
Flask-SQLAlchemy
Flask-WTF
infinity
intervals
itsdangerous
Jinja2
MarkupSafe
PyMySQL
python-dotenv
six
SQLAlchemy
validators
Werkzeug
WTForms
WTForms-Components
